Start of Fhir Query Generation refactor
